Add almost working recaptcha
Now the server needs to have cors header set - and it should be
finished!
